[Easy] Move MarketResults into its own file
